requirements.txt
annotated-types==0.7.0
anyio==4.8.0
autotiktokenizer==0.2.2
beautifulsoup4==4.12.3
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
chonkie==0.4.1
clerk-backend-api==1.7.0
click==8.1.8
colorama==0.4.6
cryptography==43.0.1
distro==1.9.0
docstring_parser==0.16
duckduckgo_search==7.2.1
eval_type_backport==0.2.2
fastapi==0.115.7
fastapi_clerk_auth==0.0.5
feedparser==6.0.11
filelock==3.17.0
fsspec==2024.12.0
gitdb==4.0.12
GitPython==3.1.44
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.27.1
idna==3.10
Jinja2==3.1.5
jiter==0.8.2
joblib==1.4.2
jsonpath-python==1.0.6
lxml==5.3.0
lxml_html_clean==0.4.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
model2vec==0.3.7
mpmath==1.3.0
mypy-extensions==1.0.0
networkx==3.4.2
newspaper4k==0.9.3.1
nltk==3.9.1
numpy==2.1.3
openai==1.60.1
packaging==24.2
pandas==2.2.3
pgvector==0.3.6
phidata==2.7.9
pillow==11.1.0
primp==0.10.1
psycopg==3.2.4
psycopg2==2.9.10
pycparser==2.22
pydantic==2.10.6
pydantic-settings==2.7.1
pydantic_core==2.27.2
Pygments==2.19.1
PyJWT==2.10.1
pypdf==5.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.20
pytz==2024.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-file==2.1.0
rich==13.9.4
safetensors==0.5.2
scikit-learn==1.6.1
scipy==1.15.1
sentence-transformers==3.4.0
setuptools==75.8.0
sgmllib3k==1.0.0
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.37
starlette==0.45.3
sympy==1.13.1
threadpoolctl==3.5.0
tiktoken==0.8.0
tldextract==5.1.3
tokenizers==0.21.0
tomli==2.2.1
torch==2.5.1
tqdm==4.67.1
transformers==4.48.1
typer==0.15.1
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
uvicorn==0.34.0